Commiting code with cleanup done to the GUI to remove all the old stats files no longer being written.  Also, modifications to the hdf5 stats file to only contain 2 parameters per size bin to define the neighborhood statistics.

+++ Source/AIM/Reconstruction/H5ReconStatsWriter.h
@@ -70,7 +70,7 @@ class AIMCOMMON_EXPORT H5ReconStatsWriter
                      Bin* eulerodf, double totalvol);
     int writeStatsData(int maxdiameter, int mindiameter,
                        double avglogdiam, double sdlogdiam, double actualgrains,
-                       const std::vector<std::vector<int> > &neighborhood,
+                       const std::vector<std::vector<double> > &neighborhood,
                        const std::vector<std::vector<double> > &svbovera,
                        const std::vector<std::vector<double> > &svcovera,
                        const std::vector<std::vector<double> > &svcoverb,
